requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
flask-sqlalchemy==2.4.3
pycryptodome==3.9.7
SQLAlchemy==1.3.17
pymysql==0.9.3
Werkzeug==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyJWT==2.3.0
requests=2.31.0